remove superflous component
