+++ packages/block-editor/src/hooks/padding.js
@@ -10,13 +10,8 @@ import {
 	useMemo,
 } from '@wordpress/element';
 import { getBlockSupport } from '@wordpress/blocks';
-import {
-	Button,
-	__experimentalUseCustomUnits as useCustomUnits,
-	__experimentalBoxControl as BoxControl,
-} from '@wordpress/components';
+import { __experimentalUseCustomUnits as useCustomUnits } from '@wordpress/components';
 import isShallowEqual from '@wordpress/is-shallow-equal';
-import { settings } from '@wordpress/icons';
 
 /**
  * Internal dependencies
@@ -30,8 +25,8 @@ import {
 } from './dimensions';
 import { cleanEmptyObject } from './utils';
 import BlockPopover from '../components/block-popover';
-import { SpacingSizeEdit, getSpacingPresetCssVar } from './spacing-size';
-
+import SpacingSizesControl from '../components/spacing-sizes-control';
+import { getSpacingPresetCssVar } from '../components/spacing-sizes-control/utils';
 /**
  * Determines if there is padding support.
  *
@@ -104,9 +99,6 @@ export function PaddingEdit( props ) {
 		setAttributes,
 	} = props;
 
-	const [ showCustomValueControl, setShowCustomValueControl ] =
-		useState( false );
-
 	const units = useCustomUnits( {
 		availableUnits: useSetting( 'spacing.units' ) || [
 			'%',
@@ -141,21 +133,8 @@ export function PaddingEdit( props ) {
 	return Platform.select( {
 		web: (
 			<>
-				{ /* <Button
-					label={
-						showCustomValueControl
-							? __( 'Use size preset' )
-							: __( 'Set custom size' )
-					}
-					icon={ settings }
-					onClick={ () => {
-						setShowCustomValueControl( ! showCustomValueControl );
-					} }
-					isPressed={ showCustomValueControl }
-					isSmall
-				/> */ }
-				{ showCustomValueControl && (
-					<BoxControl
+				<>
+					<SpacingSizesControl
 						values={ style?.spacing?.padding }
 						onChange={ onChange }
 						label={ __( 'Padding' ) }
@@ -164,31 +143,7 @@ export function PaddingEdit( props ) {
 						allowReset={ false }
 						splitOnAxis={ splitOnAxis }
 					/>
-				) }
-				{ ! showCustomValueControl && (
-					<>
-						<SpacingSizeEdit
-							values={ style?.spacing?.padding }
-							onChange={ onChange }
-							label={ __( 'Padding' ) }
-							sides={ sides }
-							units={ units }
-							allowReset={ false }
-							splitOnAxis={ splitOnAxis }
-							useSelect={ false }
-						/>
-						<SpacingSizeEdit
-							values={ style?.spacing?.padding }
-							onChange={ onChange }
-							label={ __( 'Padding' ) }
-							sides={ sides }
-							units={ units }
-							allowReset={ false }
-							splitOnAxis={ splitOnAxis }
-							useSelect={ true }
-						/>
-					</>
-				) }
+				</>
 			</>
 		),
 		native: null,
